get commands from socket
