+++ Assets/Scripts/BotSocket.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43f691de511de4bf8822013189129ef8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
